+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed oredict crafting patterns redefining recipes (Darkere)
 - Fixed Portable Grids not keeping their enchantments when placed (Darkere)
 - Shortened crafting text for the Russion translation to fix Grid overlays (yaroslav4167)
+- Fixed JEI hotkeys not working on fluid filter slots (raoulvdberge)
 
 ### 1.6.14
 - Fixed server crash (raoulvdberge)